added readme, updated package name, changed provider component from ProvideCA to CAProvider, moved useSendTransaction to its own file
